re-write newsApi to clean architecture
