+++ app/build.gradle
@@ -85,6 +85,6 @@ dependencies {
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.3.0'
 
     //Dagger
-    implementation 'com.google.dagger:dagger:2.22'
-    annotationProcessor 'com.google.dagger:dagger-compiler:2.22'
+    implementation 'com.google.dagger:dagger:2.29'
+    annotationProcessor 'com.google.dagger:dagger-compiler:2.29'
 }
