+++ app/src/main/java/by/andersen/intern/dobrov/mynewsapi/data/local/Local.java
@@ -1,0 +1,18 @@
+package by.andersen.intern.dobrov.mynewsapi.data.local;
+
+import androidx.annotation.NonNull;
+
+import java.util.List;
+
+import by.andersen.intern.dobrov.mynewsapi.domain.model.Article;
+
+public interface Local {
+
+    void getNews();
+
+    void insertNewsArticles(@NonNull List<Article> articles);
+
+    void deleteAllNewsArticles();
+
+    void getCountRows();
+}
