+++ app/src/main/java/by/andersen/intern/dobrov/mynewsapi/data/remote/Remote.java
@@ -1,0 +1,9 @@
+package by.andersen.intern.dobrov.mynewsapi.data.remote;
+
+import androidx.annotation.NonNull;
+
+public interface Remote {
+
+    void requestNews(@NonNull String keyword);
+
+}
