+++ app/src/main/java/by/andersen/intern/dobrov/mynewsapi/domain/ConnectionRepository.java
@@ -1,0 +1,9 @@
+package by.andersen.intern.dobrov.mynewsapi.domain;
+
+import androidx.annotation.NonNull;
+
+public interface ConnectionRepository {
+
+    void loadNews(@NonNull String keyword);
+
+}
