+++ app/src/main/java/by/andersen/intern/dobrov/mynewsapi/domain/model/Article.java
@@ -1,0 +1,184 @@
+package by.andersen.intern.dobrov.mynewsapi.domain.model;
+
+
+import android.os.Parcel;
+import android.os.Parcelable;
+
+import androidx.annotation.NonNull;
+import androidx.room.Entity;
+import androidx.room.PrimaryKey;
+import androidx.room.TypeConverters;
+
+import com.google.gson.annotations.Expose;
+import com.google.gson.annotations.SerializedName;
+
+import by.andersen.intern.dobrov.mynewsapi.domain.model.converter.ConverterMapper;
+
+
+@Entity(tableName = "articles")
+@TypeConverters(value = {ConverterMapper.class})
+public class Article implements Parcelable {
+
+    @PrimaryKey(autoGenerate = true)
+    private int id;
+
+    @SerializedName("source")
+    @Expose
+    private Source source;
+
+    @SerializedName("author")
+    @Expose
+    private String author;
+
+    @SerializedName("title")
+    @Expose
+    private String title;
+
+    @SerializedName("description")
+    @Expose
+    private String description;
+
+    @SerializedName("url")
+    @Expose
+    private String url;
+
+    @SerializedName("urlToImage")
+    @Expose
+    private String urlToImage;
+
+    @SerializedName("publishedAt")
+    @Expose
+    private String publishedAt;
+
+    @SerializedName("content")
+    @Expose
+    private String content;
+
+    public Article(Source source, String author, String title, String description,
+                   String url, String urlToImage, String publishedAt, String content) {
+        this.source = source;
+        this.author = author;
+        this.title = title;
+        this.description = description;
+        this.url = url;
+        this.urlToImage = urlToImage;
+        this.publishedAt = publishedAt;
+        this.content = content;
+    }
+
+    @NonNull
+    public String getContent() {
+        return content;
+    }
+
+    @NonNull
+    public Source getSource() {
+        return source;
+    }
+
+    @NonNull
+    public String getAuthor() {
+        return author;
+    }
+
+    @NonNull
+    public String getTitle() {
+        return title;
+    }
+
+    @NonNull
+    public String getDescription() {
+        return description;
+    }
+
+    @NonNull
+    public String getUrl() {
+        return url;
+    }
+
+    @NonNull
+    public String getUrlToImage() {
+        return urlToImage;
+    }
+
+    @NonNull
+    public String getPublishedAt() {
+        return publishedAt;
+    }
+
+    @NonNull
+    public int getId() {
+        return id;
+    }
+
+    @NonNull
+    public void setId(int id) {
+        this.id = id;
+    }
+
+    public void setTitle(String title) {
+        this.title = title;
+    }
+
+    public void setSource(Source source) {
+        this.source = source;
+    }
+
+    public void setPublishedAt(String publishedAt) {
+        this.publishedAt = publishedAt;
+    }
+
+    @Override
+    public String toString() {
+        return "Article{" +
+                "source=" + source +
+                ", author='" + author + '\'' +
+                ", title='" + title + '\'' +
+                ", description='" + description + '\'' +
+                ", url='" + url + '\'' +
+                ", urlToImage='" + urlToImage + '\'' +
+                ", publishedAt='" + publishedAt + '\'' +
+                ", content='" + content + '\'' +
+                '}';
+    }
+
+    @Override
+    public int describeContents() {
+        return 0;
+    }
+
+    @Override
+    public void writeToParcel(Parcel parcel, int i) {
+        parcel.writeValue(this.source);
+        parcel.writeString(this.author);
+        parcel.writeString(this.title);
+        parcel.writeString(this.description);
+        parcel.writeString(this.url);
+        parcel.writeString(this.urlToImage);
+        parcel.writeString(this.publishedAt);
+        parcel.writeString(this.content);
+    }
+
+    protected Article(Parcel in) {
+        source = (Source) in.readValue(getClass().getClassLoader());
+        author = in.readString();
+        title = in.readString();
+        description = in.readString();
+        url = in.readString();
+        urlToImage = in.readString();
+        publishedAt = in.readString();
+        content = in.readString();
+    }
+
+    public static final Creator<Article> CREATOR = new Creator<Article>() {
+        @Override
+        public Article createFromParcel(Parcel in) {
+            return new Article(in);
+        }
+
+        @Override
+        public Article[] newArray(int size) {
+            return new Article[size];
+        }
+    };
+}
