+++ app/src/main/java/by/andersen/intern/dobrov/mynewsapi/domain/model/News.java
@@ -1,0 +1,45 @@
+package by.andersen.intern.dobrov.mynewsapi.domain.model;
+
+import com.google.gson.annotations.Expose;
+import com.google.gson.annotations.SerializedName;
+
+import java.util.List;
+
+public class News {
+
+    @SerializedName("status")
+    @Expose
+    private String status;
+
+    @SerializedName("totalResult")
+    @Expose
+    private int totalResult;
+
+    @SerializedName("articles")
+    @Expose
+    private List<Article> article;
+
+    public String getStatus() {
+        return status;
+    }
+
+    public void setStatus(String status) {
+        this.status = status;
+    }
+
+    public int getTotalResult() {
+        return totalResult;
+    }
+
+    public void setTotalResult(int totalResult) {
+        this.totalResult = totalResult;
+    }
+
+    public List<Article> getArticle() {
+        return article;
+    }
+
+    public void setArticle(List<Article> article) {
+        this.article = article;
+    }
+}
