+++ app/src/main/java/by/andersen/intern/dobrov/mynewsapi/presentation/detail/adapter/GlideDataBindingAdapter.java
@@ -1,0 +1,22 @@
+package by.andersen.intern.dobrov.mynewsapi.presentation.detail.adapter;
+
+import androidx.databinding.BindingAdapter;
+import android.widget.ImageView;
+
+import com.bumptech.glide.Glide;
+import com.bumptech.glide.request.RequestOptions;
+
+public class GlideDataBindingAdapter {
+
+    @BindingAdapter("imageUrl")
+    public static void showImage(ImageView imageView, String url) {
+        RequestOptions requestOptions = new RequestOptions();
+
+        Glide
+                .with(imageView.getContext())
+                .load(url)
+                .apply(requestOptions)
+                .into(imageView);
+
+    }
+}
