+++ app/src/main/java/by/andersen/intern/dobrov/mynewsapi/presentation/list/viewmodel/SingleLiveEventForInternet.java
@@ -1,0 +1,41 @@
+package by.andersen.intern.dobrov.mynewsapi.presentation.list.viewmodel;
+
+import android.util.Log;
+
+import androidx.annotation.MainThread;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.lifecycle.LifecycleOwner;
+import androidx.lifecycle.MutableLiveData;
+import androidx.lifecycle.Observer;
+
+import java.util.concurrent.atomic.AtomicBoolean;
+
+public class SingleLiveEventForInternet<T> extends MutableLiveData<T> {
+
+    private static final String TAG = "SingleLiveEvent";
+
+    private final AtomicBoolean pending = new AtomicBoolean(false);
+
+    @MainThread
+    public void observe(@NonNull LifecycleOwner owner, @NonNull final Observer<? super T> observer) {
+
+        if (hasActiveObservers()) {
+            Log.w(TAG, "Multiple observers registered but only one will be notified of changes.");
+        }
+
+        super.observe(owner, t -> {
+            if (pending.compareAndSet(true, false)) {
+                observer.onChanged(t);
+            }
+        });
+    }
+
+
+    @MainThread
+    public void setValue(@Nullable T t) {
+        pending.set(true);
+        super.setValue(t);
+    }
+
+}
