--- app/src/main/java/by/andersen/intern/dobrov/mynewsapi/repository/ConnectionRepositoryCallback.java
@@ -1,10 +1,0 @@
-package by.andersen.intern.dobrov.mynewsapi.repository;
-
-import java.util.List;
-
-import by.andersen.intern.dobrov.mynewsapi.model.Article;
-
-public interface ConnectionRepositoryCallback {
-
-    void setArticles(List<Article> articles);
-}
