--- app/src/main/java/by/andersen/intern/dobrov/mynewsapi/repository/ConnectionRepositoryInterface.java
@@ -1,9 +1,0 @@
-package by.andersen.intern.dobrov.mynewsapi.repository;
-
-import androidx.annotation.NonNull;
-
-public interface ConnectionRepositoryInterface {
-
-    void loadNews(@NonNull String keyword);
-
-}
