+++ app/src/main/java/by/andersen/intern/dobrov/mynewsapi/util/GlobalOnlineCheck.java
@@ -1,0 +1,17 @@
+package by.andersen.intern.dobrov.mynewsapi.util;
+
+import android.content.Context;
+import android.net.ConnectivityManager;
+import android.net.NetworkInfo;
+
+public class GlobalOnlineCheck {
+
+    public static boolean isOnline(Context context) {
+        ConnectivityManager cm = (ConnectivityManager) context.getSystemService(Context.CONNECTIVITY_SERVICE);
+        NetworkInfo netInfo = cm.getActiveNetworkInfo();
+        if (netInfo != null && netInfo.isConnectedOrConnecting()) {
+            return true;
+        }
+        return false;
+    }
+}
