+++ app/src/main/res/layout/activity_main.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/colorBackground"
-    tools:context=".activity.NewsListActivity">
+    tools:context=".presentation.list.NewsListActivity">
 
 
     <androidx.swiperefreshlayout.widget.SwipeRefreshLayout
